Add poll_time aside from metric timestamp and improve overflow management
Now the poll_time take into consideration by slurmd to inquiry the driver is
the last poll time instead of the last metric timestamp. The BMC seems to grab
metrics out-of-sync.

Improved also overflow management but there are still integer issues on a task
that lose communication with slurmd, and I don't know when the BMC counter does
overflow.